add support for position args in background shorthand and protect path names containing 'left' or 'right'
